working robust navigation

+++ app/src/main/java/com/vyw/rephotoandroid/OpenCVNative.java
@@ -1,6 +1,8 @@
 package com.vyw.rephotoandroid;
 
 
+import org.opencv.core.Mat;
+
 /**
  * Created by acervenka2 on 21.04.2017.
  */
@@ -20,7 +22,7 @@ public class OpenCVNative {
     public static native float[] registration_register_point(double x, double y);
     public static native float[] registration_next_point();
     public static native int navigation_init();
-    public static native int process_navigation(long mat_current_frame, int count_frames);
+    public static native long process_navigation(long mat_current_frame, int count_frames);
 
     public static native float[] nextPoint();
